Security Config 구조 변경 및 설정 추가
