+++ src/main/java/com/myboard/config/security/SecurityConfig.java
@@ -1,58 +1,75 @@
 package com.myboard.config.security;
 
-import com.myboard.util.filter.JwtFilter;
-import com.myboard.util.filter.LoginFilter;
-import org.springframework.beans.factory.annotation.Qualifier;
+import com.myboard.exception.user.UserNotFoundException;
+import com.myboard.repository.user.UserRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Lazy;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.ProviderManager;
+import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.provisioning.InMemoryUserDetailsManager;
-import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
+@Configuration
 @EnableWebSecurity
 @EnableMethodSecurity
+@RequiredArgsConstructor
 public class SecurityConfig {
 
-    private final LoginFilter loginFilter;
-    private final JwtFilter jwtFilter;
-    private final AuthenticationEntryPoint authenticationEntryPoint;
-    private final UserDetailsService userDetailsService;
-    private final AuthenticationSuccessHandler authenticationSuccessHandler;
+    private final JwtAuthenticationFilter jwtAuthFilter;
+    private final AuthenticationProvider authenticationProvider;
+    private final LogoutHandler logoutHandler;
+    private final UserRepository userRepository;
 
-    public SecurityConfig(@Lazy LoginFilter loginFilter,
-                          JwtFilter jwtFilter,
-                          @Qualifier("customAuthenticationEntryPoint") AuthenticationEntryPoint authenticationEntryPoint,
-                          @Qualifier("customUserDetailsService") UserDetailsService userDetailsService,
-                          @Qualifier("customAuthenticationSuccessHandler") AuthenticationSuccessHandler authenticationSuccessHandler) {
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity httpSecurity) throws Exception {
+        httpSecurity
+                .csrf(AbstractHttpConfigurer::disable)
+                .authorizeHttpRequests()
+                .antMatchers("/api/v1/user/create", "/api/v1/login")
+                .permitAll()
+                .anyRequest()
+                .authenticated();
 
-        this.loginFilter = loginFilter;
-        this.jwtFilter = jwtFilter;
-        this.authenticationEntryPoint = authenticationEntryPoint;
-        this.userDetailsService = userDetailsService;
-        this.authenticationSuccessHandler = authenticationSuccessHandler;
+        httpSecurity
+                .sessionManagement()
+                .sessionCreationPolicy(SessionCreationPolicy.STATELESS);
+
+        httpSecurity
+                .authenticationProvider(authenticationProvider)
+                .addFilterBefore(jwtAuthFilter, UsernamePasswordAuthenticationFilter.class);
+
+        httpSecurity
+                .logout()
+                .logoutUrl("/api/v1/logout")
+                .addLogoutHandler(logoutHandler)
+                // 로그아웃시 로그인 정보 초기화
+                .logoutSuccessHandler((request, response, authentication) -> SecurityContextHolder.clearContext());
+
+        return httpSecurity.build();
     }
 
     @Bean
-    public UserDetailsService userDetailsService() {
+    public UserDetailsService inMemoryUserDetailsService() {
         InMemoryUserDetailsManager inMemoryUser = new InMemoryUserDetailsManager();
 
         inMemoryUser.createUser(
                 User.builder()
                         .username("user1")
-                        .password(PasswordEncoder().encode("password"))
+                        .password(passwordEncoder().encode("password"))
                         .roles("USER")
                         .build()
         );
@@ -60,15 +77,7 @@ public class SecurityConfig {
         inMemoryUser.createUser(
                 User.builder()
                         .username("user2")
-                        .password(PasswordEncoder().encode("password"))
-                        .roles("USER")
-                        .build()
-        );
-
-        inMemoryUser.createUser(
-                User.builder()
-                        .username("user3")
-                        .password(PasswordEncoder().encode("password"))
+                        .password(passwordEncoder().encode("password"))
                         .roles("USER")
                         .build()
         );
@@ -76,51 +85,35 @@ public class SecurityConfig {
         inMemoryUser.createUser(
                 User.builder()
                         .username("admin")
-                        .password(PasswordEncoder().encode("password"))
+                        .password(passwordEncoder().encode("password"))
                         .roles("ADMIN")
                         .build()
         );
-        
+
         return inMemoryUser;
     }
 
     @Bean
-    public AuthenticationManager authenticationManager() {
-        DaoAuthenticationProvider authenticationProvider = new DaoAuthenticationProvider();
-        authenticationProvider.setUserDetailsService(userDetailsService);
-        authenticationProvider.setPasswordEncoder(PasswordEncoder());
-
-        return new ProviderManager(authenticationProvider);
+    public AuthenticationProvider authenticationProvider() {
+        DaoAuthenticationProvider authProvider = new DaoAuthenticationProvider();
+        authProvider.setUserDetailsService(userDetailsService());
+        authProvider.setPasswordEncoder(passwordEncoder());
+        return authProvider;
     }
 
     @Bean
-    public SecurityFilterChain filterChain(HttpSecurity httpSecurity) throws Exception {
-        httpSecurity.csrf(csrf -> csrf.disable());
-        httpSecurity
-                .authorizeRequests()
-                .antMatchers("/api/v1/user/create", "/api/v1/login")
-                .permitAll()
-                .anyRequest()
-                .authenticated();
-        httpSecurity.addFilterAt(loginFilter, BasicAuthenticationFilter.class);
-        httpSecurity.addFilterAt(jwtFilter, BasicAuthenticationFilter.class);
-        httpSecurity
-                .sessionManagement()
-                .sessionCreationPolicy(SessionCreationPolicy.STATELESS);
-        httpSecurity
-                .exceptionHandling()
-                .authenticationEntryPoint(authenticationEntryPoint);
-
-        httpSecurity
-                .formLogin()
-                .successHandler(authenticationSuccessHandler);
-
-        return httpSecurity.build();
+    public AuthenticationManager authenticationManager(AuthenticationConfiguration config) throws Exception {
+        return config.getAuthenticationManager();
     }
 
     @Bean
-    public PasswordEncoder PasswordEncoder() {
+    public UserDetailsService userDetailsService() {
+        return username -> userRepository.findUserByUsername(username)
+                .orElseThrow(UserNotFoundException::new);
+    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
         return new BCryptPasswordEncoder();
     }
-
 }
